Refactor nystrom + tests into object oriented class structure
